Dr.SUN/requirements.txt
Django==2.1.2
djangorestframework==3.8.2
pytz==2018.5
whitenoise
django-rest-auth
django-allauth
